PCHR-1080 - Refactoring changes in unit test and BAO classes.
